MiniMart Application/requirements.txt
streamlit
pandas
mysql-connector-python
sqlalchemy
numpy
matplotlib